+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.6.2] - UNRELEASED
 ### Added
+- Global options:
+  - `colored_tht_resistors` to disable the 3D colored resistors.
+  - `field_tolerance` field/s to look for resistor tolerance.
+  - `default_resistor_tolerance` which tolerance to use when none found.
 - 3D: colored 3D models for THT resistors
 
 ### Fixed